Changed type of gov bond amount from int to long
This allows the gov to issue more than 2 trillion bonds

+++ src/jmab/agents/BondDemander.java
@@ -28,7 +28,7 @@ public interface BondDemander extends MacroAgent {
 	 * @param issuer 
 	 * @return
 	 */
-	int getBondsDemand(double price, BondSupplier issuer);
+	long getBondsDemand(double price, BondSupplier issuer);
 
 	/**
 	 * @param idBondSM
